Ctype (#283)
* Wrong Printf version requirement printing

* Small profile fixes

* Fixed crash when debug is enabled

* Semaphores are replaced by Mutexes when possible
vfprintf now use an internal optimized version of __putc_unlocked that doesn't call clib4 every time

* Reworked ctype implementation using newlib one

* don't call check_abort in stdio functions that are calling it already into hook

* ctype functions was using wrong _P define (while ctype inline #define the correct one)

+++ library/math/s_floor.c
@@ -43,7 +43,7 @@ floor(double x) {
                 if (j0 == 20) i0 += 1;
                 else {
                     j = i1 + (1 << (52 - j0));
-                    if (j < i1) i0 += 1;    /* got a carry */
+                    if (j < (uint32_t) i1) i0 += 1;    /* got a carry */
                     i1 = j;
                 }
             }
